charger-availability/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
SQLAlchemy==2.0.25
geographiclib==2.0
geopy==2.4.1
python-dotenv==1.0.1
requests==2.25.1
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
pika==1.3.2